Update JAMF-API-Get_Data_From-JAMF_Desplay_in_Excel .py
Added a try to check if the needed libraries were on the mac for Requests and Pandas. This will try to import and if not available then install with pip3.

Python3 is all that is required to be installed to run now.

You can install requests library and pandas before running the script if you want to. This just makes sure they are there when running the script.